Minor updates
Most removed configuration options were removed because they are now the default in VS Code and so they don't even need to be added. I also changed the sorting a bit so that related settings are better grouped together. Also enabled the new bread crumbs feature that was added in the new 1.26.0 release.